feat: Add PokemonGrid component to display a grid of Pokemon cards
This commit adds the PokemonGrid component, which takes an array of SimplePokemon objects as props and renders a grid of PokemonCard components. Each PokemonCard component displays information about a specific Pokemon. This component will be used in the PokemonsPage component to display a grid of Pokemon cards fetched from the PokeAPI.