feat: Add PokemonCard component to display individual Pokemon information
This commit adds the PokemonCard component, which is responsible for rendering information about a specific Pokemon. It takes a SimplePokemon object as props and displays the Pokemon's name, image, and a link to view more details. This component will be used in various parts of the application to display Pokemon information.